refactor: change method for generate data
